norma + %d

+++ data_types.c
@@ -12,7 +12,7 @@ static	long long	ft_get_type(va_list ap, t_lists *list)
 	else if ((ft_strchr(&list->mod, 'G')))
 		return (va_arg(ap, unsigned int));
 	else if ((ft_strchr(&list->mod, 'K')))
-		return ((unsigned long long)va_arg(ap, long int));
+		return ((unsigned long long)va_arg(ap, long long));
 	return (0);
 }
 
@@ -42,3 +42,4 @@ int					data_types(char *argv, va_list ap, t_lists *list)
 		ft_printf_f(list, va_arg(ap, double));
 	return (1);
 }
+
